refactor(core): extract path.rs from main.rs
Pure code move: 6 path validation / canonicalization functions
relocate from main.rs into a new core/src/path.rs. No logic change.
Function bodies are byte-identical to what they replaced.

Moved:
  canonicalize_path        — /foo -> home/foo, namespace prefix policy
  valid_world_name         — bool wrapper (CoAP surface, top-level reject)
  validate_world_name      — Result<(), reason> (HTTP 400 with reason)
  is_dot_segment           — rejects ., .., %2e, %2e%2e, mixed
  strip_dot_token          — helper for is_dot_segment
  is_reserved_world_name   — namespace roots + /proc/* subtree

All marked pub(crate). main.rs adds `mod path;` and one line to the
existing crate-root re-export block (`pub(crate) use crate::path::*;`)
so existing call sites in coap.rs (which references
`crate::canonicalize_path` / `crate::valid_world_name`) and the
inline tests in main.rs (which call them via `super::*`) keep working
with no import churn.

This is PR 2 of the v7 FSM-pipeline cascade. Base branch is
refactor/response, not master, so the diff shows only the path
extraction and not the response.rs work from PR 1.

Stats:
  120 insertions (path.rs + 1-line main.rs structural)
  85 deletions (old function bodies removed from main.rs)
  Pure mv: every inserted body is byte-identical to a deletion.

cargo fmt --manifest-path core/Cargo.toml -- --check: clean
cargo clippy --manifest-path core/Cargo.toml --all-targets -- -D warnings: clean
cargo test --manifest-path core/Cargo.toml: 100 passed, 0 failed
